backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
xgboost==2.1.1
tensorflow==2.18.0
yfinance==0.2.48
ta==0.11.0
httpx==0.27.2
python-dotenv==1.0.1
supabase==2.9.1
